+++ src/App.js
@@ -2,9 +2,11 @@ import { useState } from "react";
 import eventsData from "./data";
 import { v1 as generateUniqueID } from "uuid";
 // import Attendees from "./Attendees";
-// import Event from "./Components/Event";
-// import Footer from "./Components/Footer";
-// import Header from "./Components/Header";
+import Event from "./Components/Event";
+import Footer from "./Components/Footer";
+import Header from "./Components/Header";
+import NewEventForm from "./Components/NewEventForm";
+import Attendees from "./Components/Attendees";
 // import NewEventForm from "./Components/NewEventForm";
 
 function App() {
@@ -70,9 +72,9 @@ function App() {
     setEvents([event, ...events]);
   }
 
-  function toggleEventAttendees() {
-    setShowAttendees(!showAttendees);
-  }
+  // function toggleEventAttendees() {
+  //   setShowAttendees(!showAttendees);
+  // }
 
   function updateEventAttendance(eventId, attendeeId) {
     const eventArray = [...events];
@@ -91,129 +93,89 @@ function App() {
   return (
     <div className="App">
       <>
-        <header>
-          <h1 className="color-change-5x">RSVP App</h1>
-        </header>
+        <Header />
       </>
       <main>
         <div className="new-event">
           <>
-            <form onSubmit={handleSubmit}>
-              <h3>Create a new event</h3>
-              <label htmlFor="name">Event name:</label>
-              <input
-                type="text"
-                id="name"
-                onChange={handleTextChange}
-                value={newEvent.name}
-              />
-
-              <label htmlFor="organizer">Organizer:</label>
-              <input
-                type="text"
-                id="organizer"
-                onChange={handleTextChange}
-                value={newEvent.organizer}
-              />
-
-              <label htmlFor="eventImage">Event image:</label>
-              <input
-                type="text"
-                id="eventImage"
-                onChange={handleTextChange}
-                value={newEvent.eventImage}
-              />
-              <label htmlFor="eventType">Event type:</label>
-              <select id="eventType" onChange={handleSelectChange}>
-                <option value=""></option>
-                <option value="Birthday">Birthday</option>
-                <option value="Anniversary">Anniversary</option>
-                <option value="Intramural Sport">Intramural Sport</option>
-                <option value="Watch Party">Watch Party</option>
-                <option value="wedding">Wedding</option>
-              </select>
-              <br />
-              <input type="submit" />
-            </form>
+            <NewEventForm
+              handleAddEvent={handleAddEvent}
+              handleSubmit={handleSubmit}
+              handleTextChange={handleTextChange}
+              newEvent={newEvent}
+              handleSelectChange={handleSelectChange}
+            />
           </>
         </div>
         <div className="events">
           <ul>
             {events.map((event) => {
               const { people: attendees } = event;
-
               return (
-                <>
-                  <li key={event.id}>
-                    <img src={event.eventImage} alt={event.name} />
-                    <h5>
-                      {event.name} {event.eventType}
-                    </h5>
-                    <br />
-                    <span>Organized by: {event.organizer} </span>
-                    <br />
-                    <>
-                      <button onClick={toggleEventAttendees}>
-                        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
-                      </button>
+                <div key={event.id}>
+                  <Event
+                    event={event}
+                    updateEventAttendance={updateEventAttendance}
+                  />
+                </div>
+                // <li key={event.id}>
+                //   <img src={event.eventImage} alt={event.name} />
+                //   <h5>
+                //     {event.name} {event.eventType}
+                //   </h5>
+                //   <br />
+                //   <span>Organized by: {event.organizer} </span>
+                //   <br />
+                //   <>
+                //     <button onClick={toggleEventAttendees}>
+                //       {!showAttendees ? "Show Attendees" : "Hide Attendees"}
+                //     </button>
 
-                      {showAttendees ? (
-                        <div className="attendees">
-                          {attendees.map((attendee, index) => (
-                            <>
-                              <div key={attendee.id} className="attendee">
-                                <p>
-                                  <img
-                                    src={attendee.avatar}
-                                    alt={attendee.firstName}
-                                  />
-                                  {"   "}
-                                  <span>
-                                    {" "}
-                                    {attendee.firstName} {attendee.lastName}{" "}
-                                  </span>
-                                </p>
-                                <p>
-                                  <button
-                                    className="clickable"
-                                    onClick={() =>
-                                      updateEventAttendance(
-                                        event.id,
-                                        attendee.id
-                                      )
-                                    }
-                                  >
-                                    Attending:
-                                  </button>
-                                  <span>
-                                    {attendee.attendance ? "✅" : "❌"}
-                                  </span>
-                                </p>
+                //     {/* {showAttendees ? (
+                //       <div className="attendees">
+                //         {attendees.map((attendee, index) => (
+                //           <>
+                //             <div key={attendee.id} className="attendee">
+                //               <p>
+                //                 <img
+                //                   src={attendee.avatar}
+                //                   alt={attendee.firstName}
+                //                 />
+                //                 {"   "}
+                //                 <span>
+                //                   {" "}
+                //                   {attendee.firstName} {attendee.lastName}{" "}
+                //                 </span>
+                //               </p>
+                //               <p>
+                //                 <button
+                //                   className="clickable"
+                //                   onClick={() =>
+                //                     updateEventAttendance(event.id, attendee.id)
+                //                   }
+                //                 >
+                //                   Attending:
+                //                 </button>
+                //                 <span>{attendee.attendance ? "✅" : "❌"}</span>
+                //               </p>
 
-                                <p>
-                                  <span>Note:</span> {attendee.note}
-                                </p>
-                              </div>
-                            </>
-                          ))}
-                        </div>
-                      ) : null}
-                    </>
-                  </li>
-                </>
+                //               <p>
+                //                 <span>Note:</span> {attendee.note}
+                //               </p>
+                //             </div>
+                //           </>
+                //         ))}
+                //       </div>
+                //     ) : null} */}
+                //   </>
+                // </li>
               );
             })}
           </ul>
         </div>
       </main>
       <>
-        <footer>
-          <ul>
-            <li>Contact</li>
-            <li>About</li>
-            <li>Legal</li>
-          </ul>
-        </footer>
+        <Footer />
       </>
     </div>
   );
