+++ src/Components/Header.js
@@ -1,3 +1,3 @@
 export default function Header() {
-  return;
+  return <header>RSVP APP</header>;
 }
